+++ packages/form/src/index.ts
@@ -1,14 +1,4 @@
 export * from './components/tj-form/tj-form';
-export * from './lib/FormRemote';
-export * from './lib/FormScope';
-export * from './lib/FormValue/FormValuePluginInterface';
-export * from './lib/FormValue/FormValuePluginRegistry';
-export * from './lib/FormValue/ValuePlugins/CheckboxValuePlugin';
-export * from './lib/FormValue/ValuePlugins/CustomFormControlValuePlugin';
-export * from './lib/FormValue/ValuePlugins/InputValuePlugin';
-export * from './lib/FormValue/ValuePlugins/RadioValuePlugin';
-export * from './lib/FormValue/ValuePlugins/SelectValuePlugin';
-export * from './lib/FormValue/ValuePlugins/TextareaValuePlugin';
 export * from './lib/Scope/FormDataContainer';
 export { createFormScope, createScope, defineArray, defineScope } from './lib/Scope/scope';
 export type {